refactor(dynamic): reorganize type description modules
