+++ crates/ros-z/src/extended_schema.rs
@@ -1,23 +1,10 @@
-use std::sync::{Arc, RwLock};
-use std::time::Duration;
+use std::sync::Arc;
 
 use serde::{Deserialize, Serialize};
 use sha2::Digest;
-use tracing::{debug, info, warn};
-use zenoh::Session;
 
-use crate::context::GlobalCounter;
 use crate::dynamic::{DynamicError, FieldType, MessageSchema};
-use crate::entity::{
-    EndpointEntity, EndpointKind, Entity, EntityKind, NodeEntity, TypeHash, TypeInfo,
-};
-use crate::graph::Graph;
-use crate::msg::{SerdeCdrSerdes, ZMessage, ZService};
 use crate::node::ZNode;
-use crate::service::{ZClient, ZClientBuilder, ZServer, ZServerBuilder};
-use crate::{Builder, ServiceTypeInfo};
-
-const EXTENDED_SERVICE_NAME: &str = "~get_extended_type_description";
 
 pub trait ExtendedMessageTypeInfo: crate::MessageTypeInfo {
     fn extended_message_schema() -> Arc<MessageSchema>;
@@ -345,413 +332,4 @@ fn extended_to_enum_payload(
             fields.into_iter().map(extended_to_field_schema).collect(),
         ),
     }
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct GetExtendedTypeDescriptionRequest {
-    pub type_name: String,
-    pub type_hash: String,
-}
-
-impl ZMessage for GetExtendedTypeDescriptionRequest {
-    type Serdes = SerdeCdrSerdes<Self>;
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct GetExtendedTypeDescriptionResponse {
-    pub successful: bool,
-    pub failure_reason: String,
-    pub type_hash: String,
-    pub schema_json: String,
-}
-
-impl ZMessage for GetExtendedTypeDescriptionResponse {
-    type Serdes = SerdeCdrSerdes<Self>;
-}
-
-pub struct GetExtendedTypeDescription;
-
-impl ZService for GetExtendedTypeDescription {
-    type Request = GetExtendedTypeDescriptionRequest;
-    type Response = GetExtendedTypeDescriptionResponse;
-}
-
-impl ServiceTypeInfo for GetExtendedTypeDescription {
-    fn service_type_info() -> TypeInfo {
-        TypeInfo::new(
-            "ros_z::srv::dds_::GetExtendedTypeDescription_",
-            TypeHash::zero(),
-        )
-    }
-}
-
-#[derive(Clone)]
-pub struct RegisteredExtendedSchema {
-    pub schema: Arc<MessageSchema>,
-    pub type_hash: String,
-}
-
-impl RegisteredExtendedSchema {
-    pub fn new(schema: Arc<MessageSchema>) -> Result<Self, DynamicError> {
-        let type_hash = compute_extended_type_hash(&schema)?.to_rihs_string();
-        Ok(Self { schema, type_hash })
-    }
-}
-
-#[derive(Clone)]
-pub struct ExtendedTypeDescriptionService {
-    schemas: Arc<RwLock<std::collections::HashMap<String, RegisteredExtendedSchema>>>,
-    _server: Arc<ZServer<GetExtendedTypeDescription, ()>>,
-}
-
-impl ExtendedTypeDescriptionService {
-    pub fn new(
-        session: Arc<Session>,
-        node_name: &str,
-        namespace: &str,
-        node_id: usize,
-        counter: &GlobalCounter,
-        clock: &crate::time::ZClock,
-    ) -> zenoh::Result<Self> {
-        let schemas = Arc::new(RwLock::new(std::collections::HashMap::new()));
-        let node_entity = NodeEntity::new(
-            0,
-            session.zid(),
-            node_id,
-            node_name.to_string(),
-            namespace.to_string(),
-            String::new(),
-        );
-
-        let entity = EndpointEntity {
-            id: counter.increment(),
-            node: Some(node_entity),
-            kind: EndpointKind::Service,
-            topic: EXTENDED_SERVICE_NAME.to_string(),
-            type_info: Some(GetExtendedTypeDescription::service_type_info()),
-            qos: Default::default(),
-        };
-
-        let server_builder: ZServerBuilder<GetExtendedTypeDescription> = ZServerBuilder {
-            entity,
-            session,
-            clock: clock.clone(),
-            keyexpr_format: ros_z_protocol::KeyExprFormat::default(),
-            _phantom_data: Default::default(),
-        };
-
-        let schemas_clone = schemas.clone();
-        let server = server_builder.build_with_callback(move |query| {
-            Self::handle_query(&schemas_clone, &query);
-        })?;
-
-        info!(
-            "[ETDS] ExtendedTypeDescriptionService created for node: {}/{}",
-            namespace, node_name
-        );
-
-        Ok(Self {
-            schemas,
-            _server: Arc::new(server),
-        })
-    }
-
-    pub fn register_schema(&self, schema: Arc<MessageSchema>) -> Result<(), DynamicError> {
-        let registered = RegisteredExtendedSchema::new(schema.clone())?;
-        let type_name = schema.type_name.clone();
-
-        let mut schemas = self
-            .schemas
-            .write()
-            .map_err(|_| DynamicError::RegistryLockPoisoned)?;
-        schemas.insert(type_name.clone(), registered);
-
-        debug!("[ETDS] Registered extended schema: {}", type_name);
-        Ok(())
-    }
-
-    pub fn get_schema(&self, type_name: &str) -> Result<Option<Arc<MessageSchema>>, DynamicError> {
-        let schemas = self
-            .schemas
-            .read()
-            .map_err(|_| DynamicError::RegistryLockPoisoned)?;
-        Ok(schemas
-            .get(type_name)
-            .map(|registered| registered.schema.clone()))
-    }
-
-    fn handle_query(
-        schemas: &Arc<RwLock<std::collections::HashMap<String, RegisteredExtendedSchema>>>,
-        query: &zenoh::query::Query,
-    ) {
-        let request = match query.payload() {
-            Some(payload) => match <GetExtendedTypeDescriptionRequest as ZMessage>::deserialize(
-                payload.to_bytes().as_ref(),
-            ) {
-                Ok(request) => request,
-                Err(err) => {
-                    warn!("[ETDS] Failed to decode request: {}", err);
-                    return;
-                }
-            },
-            None => {
-                warn!("[ETDS] Missing request payload");
-                return;
-            }
-        };
-
-        let response = Self::build_response(schemas, &request);
-        let bytes = ZMessage::serialize(&response);
-        use zenoh::Wait;
-        if let Err(err) = query.reply(query.key_expr().clone(), bytes).wait() {
-            warn!("[ETDS] Failed to send response: {}", err);
-        }
-    }
-
-    fn build_response(
-        schemas: &Arc<RwLock<std::collections::HashMap<String, RegisteredExtendedSchema>>>,
-        request: &GetExtendedTypeDescriptionRequest,
-    ) -> GetExtendedTypeDescriptionResponse {
-        let schemas_guard = match schemas.read() {
-            Ok(guard) => guard,
-            Err(_) => {
-                return GetExtendedTypeDescriptionResponse {
-                    successful: false,
-                    failure_reason: "Internal error: registry lock poisoned".to_string(),
-                    type_hash: String::new(),
-                    schema_json: String::new(),
-                };
-            }
-        };
-
-        let registered = match schemas_guard.get(&request.type_name) {
-            Some(schema) => schema,
-            None => {
-                return GetExtendedTypeDescriptionResponse {
-                    successful: false,
-                    failure_reason: format!("Type '{}' not registered", request.type_name),
-                    type_hash: String::new(),
-                    schema_json: String::new(),
-                };
-            }
-        };
-
-        if !request.type_hash.is_empty() && request.type_hash != registered.type_hash {
-            return GetExtendedTypeDescriptionResponse {
-                successful: false,
-                failure_reason: format!(
-                    "Type hash mismatch: expected {}, got {}",
-                    registered.type_hash, request.type_hash
-                ),
-                type_hash: String::new(),
-                schema_json: String::new(),
-            };
-        }
-
-        match schema_to_extension_json(&registered.schema) {
-            Ok(schema_json) => GetExtendedTypeDescriptionResponse {
-                successful: true,
-                failure_reason: String::new(),
-                type_hash: registered.type_hash.clone(),
-                schema_json,
-            },
-            Err(err) => GetExtendedTypeDescriptionResponse {
-                successful: false,
-                failure_reason: format!("Failed to serialize extended schema: {}", err),
-                type_hash: String::new(),
-                schema_json: String::new(),
-            },
-        }
-    }
-}
-
-pub struct ExtendedTypeDescriptionClient {
-    session: Arc<Session>,
-    counter: Arc<GlobalCounter>,
-    graph: Option<Arc<Graph>>,
-    timeout: Duration,
-}
-
-impl ExtendedTypeDescriptionClient {
-    pub fn new(session: Arc<Session>, counter: Arc<GlobalCounter>) -> Self {
-        Self {
-            session,
-            counter,
-            graph: None,
-            timeout: Duration::from_secs(10),
-        }
-    }
-
-    pub fn with_graph(
-        session: Arc<Session>,
-        counter: Arc<GlobalCounter>,
-        graph: Arc<Graph>,
-    ) -> Self {
-        Self {
-            session,
-            counter,
-            graph: Some(graph),
-            timeout: Duration::from_secs(10),
-        }
-    }
-
-    pub fn with_timeout(mut self, timeout: Duration) -> Self {
-        self.timeout = timeout;
-        self
-    }
-
-    pub async fn get_type_description(
-        &self,
-        node_name: &str,
-        namespace: &str,
-        type_name: &str,
-        type_hash: &str,
-    ) -> Result<GetExtendedTypeDescriptionResponse, DynamicError> {
-        let service_name = if namespace.is_empty() || namespace == "/" {
-            format!("/{}/get_extended_type_description", node_name)
-        } else {
-            format!("{}/{}/get_extended_type_description", namespace, node_name)
-        };
-
-        let client = self.create_client(&service_name, "")?;
-        self.query_with_client(&client, type_name, type_hash).await
-    }
-
-    async fn query_with_client(
-        &self,
-        client: &ZClient<GetExtendedTypeDescription>,
-        type_name: &str,
-        type_hash: &str,
-    ) -> Result<GetExtendedTypeDescriptionResponse, DynamicError> {
-        let request = GetExtendedTypeDescriptionRequest {
-            type_name: type_name.to_string(),
-            type_hash: type_hash.to_string(),
-        };
-
-        client
-            .call_or_timeout(&request, self.timeout)
-            .await
-            .map_err(|e| DynamicError::SerializationError(e.to_string()))
-    }
-
-    pub async fn get_type_description_for_topic(
-        &self,
-        topic: &str,
-        timeout: Duration,
-    ) -> Result<(Arc<MessageSchema>, String), DynamicError> {
-        let graph = self.graph.as_ref().ok_or_else(|| {
-            DynamicError::SerializationError(
-                "ExtendedTypeDescriptionClient requires graph for topic-based discovery"
-                    .to_string(),
-            )
-        })?;
-
-        let start = tokio::time::Instant::now();
-        let mut last_error = None;
-
-        while start.elapsed() < timeout {
-            let publishers = graph.get_entities_by_topic(EntityKind::Publisher, topic);
-            if publishers.is_empty() {
-                tokio::time::sleep(Duration::from_millis(250)).await;
-                continue;
-            }
-
-            for publisher in &publishers {
-                let Entity::Endpoint(endpoint) = &**publisher else {
-                    continue;
-                };
-
-                let type_info = match &endpoint.type_info {
-                    Some(info) => info,
-                    None => continue,
-                };
-                let Some(node) = endpoint.node.as_ref() else {
-                    last_error = Some(DynamicError::MissingNodeIdentity {
-                        topic: topic.to_string(),
-                    });
-                    continue;
-                };
-                let type_name = normalize_type_name(&type_info.name);
-                let type_hash = type_info.hash.to_rihs_string();
-
-                match self
-                    .get_type_description(&node.name, &node.namespace, &type_name, &type_hash)
-                    .await
-                {
-                    Ok(response) => {
-                        let schema = Self::response_to_schema(&response)?;
-                        return Ok((schema, response.type_hash));
-                    }
-                    Err(err) => last_error = Some(err),
-                }
-            }
-
-            tokio::time::sleep(Duration::from_millis(250)).await;
-        }
-
-        Err(last_error.unwrap_or_else(|| {
-            DynamicError::SchemaNotFound(format!(
-                "Failed to get extended type description from any publisher on {}",
-                topic
-            ))
-        }))
-    }
-
-    pub fn response_to_schema(
-        response: &GetExtendedTypeDescriptionResponse,
-    ) -> Result<Arc<MessageSchema>, DynamicError> {
-        if !response.successful {
-            return Err(DynamicError::SerializationError(format!(
-                "Response indicates failure: {}",
-                response.failure_reason
-            )));
-        }
-
-        schema_from_extension_json(&response.schema_json)
-    }
-
-    fn create_client(
-        &self,
-        service_name: &str,
-        namespace: &str,
-    ) -> Result<ZClient<GetExtendedTypeDescription>, DynamicError> {
-        let node_entity = NodeEntity::new(
-            0,
-            self.session.zid(),
-            self.counter.increment(),
-            "extended_type_desc_client".to_string(),
-            namespace.to_string(),
-            String::new(),
-        );
-
-        let entity = EndpointEntity {
-            id: self.counter.increment(),
-            node: Some(node_entity),
-            kind: EndpointKind::Client,
-            topic: service_name.to_string(),
-            type_info: Some(GetExtendedTypeDescription::service_type_info()),
-            qos: Default::default(),
-        };
-
-        let builder: ZClientBuilder<GetExtendedTypeDescription> = ZClientBuilder {
-            entity,
-            session: self.session.clone(),
-            clock: crate::time::ZClock::default(),
-            keyexpr_format: ros_z_protocol::KeyExprFormat::default(),
-            _phantom_data: Default::default(),
-        };
-
-        builder
-            .build()
-            .map_err(|e| DynamicError::SerializationError(e.to_string()))
-    }
-}
-
-fn normalize_type_name(dds_name: &str) -> String {
-    dds_name
-        .replace("::msg::dds_::", "/msg/")
-        .replace("::srv::dds_::", "/srv/")
-        .replace("::action::dds_::", "/action/")
-        .trim_end_matches('_')
-        .to_string()
 }
